python/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# YOLOv8 and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Data Validation
pydantic>=2.0.0

# Additional Dependencies
PyYAML>=6.0
requests>=2.31.0